rock_paper_scissor repo
Its a simple game called Rock Paper Scissor. This games rule is if Rock bit scissor then rock win, if paper hold the rock then paper wins and lastly if scissors cut the paper then scissors then scissors win. The game is designed with simple if- else operator and by using of while loop it is easy to design.